Append "*." prefix to all ingress domain names
Domain names listed for all ingress routes in ClusterDeployment spec need wildcard certificates. However, the domain names do not have "*." prefix to them. When a CertificaeRequest is created for a cluster, we must add the prefix so wildcard certificates are requested for that domain.

Ingress domain `foo.bar.com` becomes `*.foo.bar.com` for the certificate request.

Signed-off-by: Tejas Parikh <tparikh@redhat.com>