Update README, remove vestigial lets_encrypt_environment
This adds a generated ToC, fixes some broken links, updates references
to older required versions of things, addresses some typographical
errors, and generally tidies things up.

Almost no meaning/content is changed.

While doing this, it was discovered that the `lets_encrypt_environment`
configmap key is unused, but had a vestigial presence in a const and
test as well as the README. These are cleaned up here.