OSD-10349 better tests (#220)
* rename ACMEClient to LetsEncryptClient

also FakeACMEClient to FakeLetsEncryptClient

this reduces confusion between:
* ACMEClient (defined in github.com/openshift/certman-operator/pkg/leclient)
* ACMEClient.Client (defined in github.com/openshift/certman-operator/pkg/leclient)
* acme.Client (defined in github.com/eggsampler/acme and is the type of
  ACMEClient.Client)

* rename Client to LetsEncryptClientInterface

since this is the interface for LetsEncryptClient (defined immediately
below), the interface name will now clearly indicate what the type is.
additionally, this commit changes functions the certificaterequest
controller to accept and return LetsEncryptClientInterface instead of
LetsEncryptClient, making testing easier

* create AcmeClientInterface interface

AcmeClientInterface is an interface for github.com/eggsampler/acme.Client
defining parameters and return values as interfaces is just good
practice but also makes testing much easier.

* add unit tests for leclient.UpdateAccount()

* add unit test for leclient.CreateOrder()

* add unit test for leclient.GetOrderURL()

also since there's no way for leclient.GetOrderURL() to generate an
error, that has been removed from the function's signature and the
relevant bits of code have been cleaned up to match it

* add unit tests for leclient.OrderAuthorization()

* add unit tests for leclient.FetchAuthorization()

* add unit tests for leclient.GetAuthorizationURL()

* add unit test for leclient.GetAuthorizationIdentifier()

* add unit tests for leclient.SetChallengeType()

also since there's no way for leclient.leclient.SetChallengeType() to generate
an error, that has been removed from the function's signature and the
relevant bits of code have been cleaned up to match it

* add unit test for leclient.GetDNS01KeyAuthorization()

* add unit test for leclient.GetChallengeURL()

* add unit test for leclient.UpdateChallenge()

* add unit tests for leclient.FinalizeOrder()

* add unit test for leclient.GetOrderEndpoint()

* add unit tests for leclient.FetchCertificates()

* add unit tests for leclient.RevokeCertificate()

* move acme client interface and mock client into dedicated package

this helps avoid import loops when trying to use one or the other

* no need for a mock leclient anymore

now that there's a mock acme client, we can just use the real leclient
and inject the mock acme client instead

* add empty test file for acmeclient

acmeclient contains nothing but an interface definition but adding the
test file lets `go test ./...` confirm the coverage anyway

* add unit tests for the mock acme client

* add functionality to the mock acme client

implements functionality:
* NewFakeAcmeClient() (and FakeAcmeClientOptions) can set more values to
  determine the mock client's behavior
* FetchCertificates() returns certificates
* FetchAuthorization() returns an authorization
* NewOrder() returns an order

* add unit tests for certificaterequest.IssueCertificate()

NB this includes a check in issue_certificate.go to see if it's being
run in a test environment and, if so, skips the
VerifyDnsResourceRecordUpdate(). this isn't ideal for a bunch of reasons
but will require heavy refactoring to avoid it

* add more empty test files for files with no statements to test

* add unit tests for client.NewClient()

* implement a mock (dns) client

* enable client.NewClient() to create a mock client

this will allow for smoother testing of the operator's controllers by
avoiding the need to reach out to platforms' DNS services, mocking the
calls instead.

the certmanv1alpha1 APIs have been expanded to include a
MockPlatformSecrets struct so the mock client can be configured from
kube objects

* unit test for the certificaterequest reconcile loop getting a new cert

use the mock acme client to return a pre-defined ssl certificate to the
certificaterequest controller

* add empty test file for version.go

this lets `go test ./...` return "ok" for
github.com/openshift/certman-operator/version instead of "?"

* add unit tests confirming controllers are added to mgr

adding the controllers to the manager are done in init() functions so
these tests confirm that the controllers' Add() function are included in
AddToManagerFuncs

* move the mock route53iface client into a dedicated package

* add test coverage for our aws dns client

* add empty test file for gcp dns client

this lets us start tracking test coverage at least

* controller-gen

* fix validation stuff

`make validate` fails on my system so i'm just copying the things the
prow job is complaining about