expanding GCP client scope to support impersonation
The GCP client used by the certman operator is instantiated with
credentials that are limited in scope to only perform dns operations.
This approach works with credentials in the form of long lived keys, but
not with short-lived impersonation credentials. To follow
best-practices, the certman operator needs to support credentials which impersonate
service accounts, as in the form shown below.
```
{
  "project_id": "PROJECT_ID",
  "delegates": [],
  "service_account_impersonation_url": "https://iamcredentials.googleapis.com/v1/projects/-/serviceAccounts/SERVICE_ACCOUNT_ID@PROJECT_ID.iam.gserviceaccount.com:generateAccessToken",
  "source_credentials": { REDACTED },
  "type": "impersonated_service_account",
  "universe_domain": "googleapis.com"
}
```

To support credentials of the above form, the scope must be expanded to
call the iam method "generateAccessToken". This is acheived simply by
adding "https://www.googleapis.com/auth/cloud-platform" to the set of
scopes to be associated with the credentials.